requirements.txt
ultralytics>=8.0.0
streamlit>=1.28.0
opencv-python>=4.8.0
playsound==1.2.2
plotly>=5.0.0
python-dotenv>=1.0.0
twilio>=8.0.0
Pillow>=9.0.0
